% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appraisal.R
\name{appraisal}
\alias{appraisal}
\title{appraisal}
\usage{
appraisal(observed, LP, plots = TRUE, title = "", weight = rep(1,
  length(observed)))
}
\arguments{
\item{observed}{Number of observations in the data set.}

\item{LP}{Calculated linear predictor for all indiviuals in the data set
based on the prediction model being validated.}

\item{plots}{Logical. If TRUE, calibration plots are returned.}

\item{title}{Optional title string for calibration plots.}

\item{weight}{Optional weights for weighted analyses.}
}
\value{
\code{appraisal} returns a matrix of performance estimates and
                         optional calibration plot
}
\description{
Estimates of prediction model performance
}
\details{
This function calculates simple measures of prediction model performance,
currently Harrell's c-index, O:E ratios and optional calibration plots.

This function is currently only compatible with logistic regression models.
}
\examples{
## Example 1: validation of a simple prediction rule. Unweighted analysis
covmat <- matrix(c(0.2,0,0,0.2), nrow=2)
set.seed(123)
d <- datafy(obs = 100, means = c(0,0), covmat = covmat,
     var.names = c("X1", "X2", "Y"), genmod = c(-1.5, 1, 1))
m <- glm(d$Y ~ d$X1 + d$X2, family=binomial)
set.seed(456)
covmat <- matrix(c(0.4,0,0,0.4), nrow=2)
v <- datafy(obs = 100, means = c(0.0,0.0), covmat = covmat,
     var.names = c("X1", "X2", "Y"), genmod = c(-2, 1.2, 1.2))
LP <- m$coef[1] + m$coef[2]*v$X1 + m$coef[3]*v$X2
appraisal(obs = v$Y, LP = LP, plots = TRUE, title = "Calibration plot")
}

