% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datafy.R
\name{datafy}
\alias{datafy}
\title{datafy}
\usage{
datafy(obs, means, covmat, var.names, genmod)
}
\arguments{
\item{obs}{Number of observations in the data set.}

\item{means}{A vector of predictor means.}

\item{covmat}{A covariance matrix (see example).}

\item{var.names}{A string of variable names (optional).}

\item{genmod}{Vector of parameter values for a logistic regression model
used in the generation of binary outcomes.}
}
\value{
\code{datafy} returns a dataframe with m predictors and a binary outcome
}
\description{
Generate a data set
}
\details{
This function simulates a data set for use in simulations of model
development, validation or updating.

Currently supports the generation of correlated continuous, normally distributed
predictor variables and a binary outcome.
}
\examples{
## Example 1: Generation of an untreated development set
covmat <- matrix(c(0.2,0,0,0.2), nrow=2)
d <- datafy(obs = 100, means = c(0,0), covmat = covmat,
     var.names = c("X1", "X2", "Y"), genmod = c(-1.5, 1, 1))
}

